+++ app/src/main/java/com/example/mvvm_template/domain/entity/Cart.kt
@@ -3,8 +3,8 @@ package com.example.mvvm_template.domain.entity
 data class Cart(
     val code: String?=null,
     val discountName: String?=null,
-    val discountValue: Int?=null,
-    val subTotal: Int?=null,
-    val total: Int?=null,
+    val discountValue: Double?=null,
+    val subTotal: Double?=null,
+    val total: Double?=null,
     val products: List<Product>?=null
 )
